Move carousel to new file
